environments_harbor/distributed_training_hang_medium/environment/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
PyYAML>=5.4
numpy>=1.19.0